+++ modules/task_3/purtova_k_shell_sort_batcher_merge/shell_sort.h
@@ -1,11 +1,10 @@
 // Copyright 2022 Purtova Ksenia
 #ifndef MODULES_TASK_3_PURTOVA_K_SHELL_SORT_BATCHER_MERGE_SHELL_SORT_H_
 #define MODULES_TASK_3_PURTOVA_K_SHELL_SORT_BATCHER_MERGE_SHELL_SORT_H_
-
 #include <vector>
 
 std::vector<int> Shell(const std::vector<int>& A, int n);
-std::vector <int> MergerEvenOdd(const std::vector<int>& arr1, const std::vector<int>& arr2, int i1, int i2, int size_result);
+std::vector <int> MergerEvenOdd(const std::vector<int>& arr1, const std::vector<int>& arr2, int i1, int i2, int size);
 std::vector<int> BatcherEvenNumber(const std::vector<int>& arr1, const std::vector<int>& arr2);
 std::vector<int> BatcherOddNumber(const std::vector<int>& arr1, const std::vector<int>& arr2);
 std::vector<int> BatcherMerger(const std::vector<int>& arr1, const std::vector<int>& arr2);
